% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{prepare_data}
\alias{prepare_data}
\title{Preparation of data}
\usage{
prepare_data(thdata_df, port_num)
}
\arguments{
\item{thdata_df}{data itself}

\item{port_num}{port number}
}
\value{
dataframe with requested port
}
\description{
Preparation of data
}
